bot/requirements.txt
scikit-learn==1.5.1
joblib==1.4.2
sqlalchemy==2.0.32
pymysql==1.1.1
python-dotenv==1.0.1
